requirements.txt
flask
python-dotenv
line-bot-sdk
apscheduler
gunicorn